Initial draft
# Salish Sea Geofence
## Description
This is a website to promote a proposal to build a prototype of software that uses a “GeoFence” to alert boaters when they are inside the eelgrass zone, a common feature of the nearshore region in the San Juan Islands, and inform them that dropping their anchor in this area is not ethical or wise. First because Native eelgrass is a foundational species that sequesters carbon and provides valued ecological, economic and cultural services that are protected by federal, state, county and Tribal governments.
and second submerged habitats in this zone do not insure a safe anchorage;
## Objective
An initial prototype application will be developed to explore several aspects of the technology, the data, the use cases, the user experience, and test the effectiveness of the combination of native eelgrass boundaries and cellular location, both online and offline as a potential tool for boater education and awareness in protecting existing or potential eelgrass zones.
## Scope of Work
The work will be a contract “work for hire” effort with the funding paying for the hourly time by the contractors to put the data and software in place. The prototype will be Open Source (license type TBD but probably MIT) and as such will set the example for further work to also be Open Source and Open Data. Eelgrass restoration efforts are underway in the Salish Sea and a heightened level of boater awareness and improved boater education will significantly enhance these efforts. The current vision is for 3 phases, the first clarifies our technical choices and draft requirements for the pilot, the second phase is the pilot data and code. Please note the end of this phase is still a “pilot” and not a complete app. or where we want the final application to be. The third phase, least defined or budgeted at this point, is a finished application for boaters covering the initial area of interest in the Salish Sea.